Move more common functions to voight kampff tools
